% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{detect_split_strips}
\alias{detect_split_strips}
\title{Detect bimodal source X distribution and compute two strip windows}
\usage{
detect_split_strips(
  src_crs,
  src_gt,
  src_dim,
  dst_crs,
  dst_gt,
  dst_off,
  dst_size
)
}
\description{
Detect bimodal source X distribution and compute two strip windows
}
\keyword{internal}
