% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overview.R
\name{get_overview_dims}
\alias{get_overview_dims}
\title{Get overview dimensions for all overview levels}
\usage{
get_overview_dims(ds, band, n_ovr, src_dim)
}
\arguments{
\item{ds}{GDALRaster object}

\item{band}{Band number}

\item{n_ovr}{Number of overviews}

\item{src_dim}{Full resolution c(ncol, nrow)}
}
\value{
Matrix with n_ovr rows and 2 columns (xsize, ysize),
ordered from finest (overview 1) to coarsest (overview n_ovr).
}
\description{
Get overview dimensions for all overview levels
}
\keyword{internal}
