% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rust_fetch_decode_tile}
\alias{rust_fetch_decode_tile}
\title{Fetch a tile via HTTP range request and decode it.}
\usage{
rust_fetch_decode_tile(
  url,
  byte_offset,
  byte_length,
  tile_width,
  tile_height,
  predictor
)
}
\arguments{
\item{url}{Character URL}

\item{byte_offset}{Numeric byte offset}

\item{byte_length}{Integer byte count}

\item{tile_width}{Integer tile width}

\item{tile_height}{Integer tile height}

\item{predictor}{Integer predictor type}
}
\value{
Integer vector of decoded pixel values
}
\description{
Fetch a tile via HTTP range request and decode it.
}
